add ability to take in necessary auth call backs
