--- src/types/index.d.ts
@@ -1,10 +1,0 @@
-export interface Vessel {
-  init: (options: object) => void;
-  open: () => void;
-}
-
-declare global {
-  interface Window {
-    Vessel: Vessel;
-  }
-}
